+++ bip32/src/extended_key/private_key.rs
@@ -8,7 +8,7 @@ use core::{
     fmt::{self, Debug},
     str::FromStr,
 };
-use hmac::Mac;
+use hmac::{KeyInit, Mac};
 use subtle::{Choice, ConstantTimeEq};
 use zeroize::Zeroize;
 
@@ -26,7 +26,6 @@ const BIP39_DOMAIN_SEPARATOR: [u8; 12] = [
 
 /// Extended private secp256k1 ECDSA signing key.
 #[cfg(feature = "secp256k1")]
-#[cfg_attr(docsrs, doc(cfg(feature = "secp256k1")))]
 pub type XPrv = ExtendedPrivateKey<k256::ecdsa::SigningKey>;
 
 /// Extended private keys derived using BIP32.
@@ -52,7 +51,6 @@ where
 
     /// Derive a child key from the given [`DerivationPath`].
     #[cfg(feature = "alloc")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
     pub fn derive_from_path<S>(seed: S, path: &DerivationPath) -> Result<Self>
     where
         S: AsRef<[u8]>,
@@ -90,23 +88,11 @@ where
     /// Derive a child key for a particular [`ChildNumber`].
     pub fn derive_child(&self, child_number: ChildNumber) -> Result<Self> {
         let depth = self.attrs.depth.checked_add(1).ok_or(Error::Depth)?;
+        let (tweak, chain_code) = self
+            .private_key
+            .derive_tweak(&self.attrs.chain_code, child_number)?;
 
-        let mut hmac =
-            HmacSha512::new_from_slice(&self.attrs.chain_code).map_err(|_| Error::Crypto)?;
-
-        if child_number.is_hardened() {
-            hmac.update(&[0]);
-            hmac.update(&self.private_key.to_bytes());
-        } else {
-            hmac.update(&self.private_key.public_key().to_bytes());
-        }
-
-        hmac.update(&child_number.to_bytes());
-
-        let result = hmac.finalize().into_bytes();
-        let (child_key, chain_code) = result.split_at(KEY_SIZE);
-
-        // We should technically loop here if a `secret_key` is zero or overflows
+        // We should technically loop here if the tweak is zero or overflows
         // the order of the underlying elliptic curve group, incrementing the
         // index, however per "Child key derivation (CKD) functions":
         // https://github.com/bitcoin/bips/blob/master/bip-0032.mediawiki#child-key-derivation-ckd-functions
@@ -115,12 +101,12 @@ where
         //
         // ...so instead, we simply return an error if this were ever to happen,
         // as the chances of it happening are vanishingly small.
-        let private_key = self.private_key.derive_child(child_key.try_into()?)?;
+        let private_key = self.private_key.derive_child(tweak)?;
 
         let attrs = ExtendedKeyAttrs {
             parent_fingerprint: self.private_key.public_key().fingerprint(),
             child_number,
-            chain_code: chain_code.try_into()?,
+            chain_code,
             depth,
         };
 
@@ -163,7 +149,6 @@ where
 
     /// Serialize this key as a self-[`Zeroizing`] `String`.
     #[cfg(feature = "alloc")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
     pub fn to_string(&self, prefix: Prefix) -> Zeroizing<String> {
         Zeroizing::new(self.to_extended_key(prefix).to_string())
     }